Renames package to have http client version
